Fixed possible double type issue (replaced with int)

+++ praktikum1/src/main/java/aufgabe1/utils/GraphUtils.java
@@ -34,7 +34,7 @@ public final class GraphUtils {
     public static Network<Vertex, NetworkEdge> convertToNetwork(Graph<Vertex, DefaultWeightedEdge> graph) {
         Network<Vertex, NetworkEdge> network = new Network();
         graph.vertexSet().forEach(vertex -> network.addVertex(vertex));
-        graph.edgeSet().forEach(edge -> network.addEdge(graph.getEdgeSource(edge), graph.getEdgeTarget(edge)).setCapacity(graph.getEdgeWeight(edge)));
+        graph.edgeSet().forEach(edge -> network.addEdge(graph.getEdgeSource(edge), graph.getEdgeTarget(edge)).setCapacity((int) graph.getEdgeWeight(edge)));
         return network;
     }
